Ignore bad points in radial velocity/continuum fit split util into SME and general part
